Renamed slider components to 'Carousel' for clarity and moved all components to their own folder.
