+++ src/components/Carousel/Carousel.js
@@ -1,12 +1,10 @@
 import React from 'react'
 import styled from '@emotion/styled'
 
-const WorkspaceSlider = styled.div`
+export const Carousel = styled.div`
   transform: translateX(-${props => props.translate}px);
   transition: transform ease-out ${props => props.transition}s;
   height: 100%;
   width: ${props => props.width}px;
   display: flex;
 `
-
-export default WorkspaceSlider
