+++ src/components/Carousel/arrowLftVec.png
