+++ src/components/Carousel/arrowRghtVec.png
